+++ src/main/java/com/kjh/clean_stock/web/dto/Receipt/ReceiptApiSaveDto.java
@@ -9,6 +9,7 @@ import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Getter
 @NoArgsConstructor
@@ -17,14 +18,14 @@ public class ReceiptApiSaveDto {
     private int stockCnt;
     @Min(value = 0 , message = "0 이상의 값을 넣어주세요.")
     @NotNull
-    private Long stockAvr;
+    private BigDecimal stockAvr;
     @NotNull
     private Long portfolio_id;
     @NotNull
     private Long stock_id;
 
     @Builder
-    public ReceiptApiSaveDto(int stockCnt, Long stockAvr, Long portfolio_id,Long stock_id){
+    public ReceiptApiSaveDto(int stockCnt, BigDecimal stockAvr, Long portfolio_id,Long stock_id){
         this.stockCnt= stockCnt;
         this.stockAvr=stockAvr;
         this.portfolio_id= portfolio_id;
